move submissions to components and make it clearer when they're advertisers
